Adding tiles, plant eating system, regrowth system, time, plant spreading system

+++ world_generator/civilizations/historicalEvent.cpp
@@ -1,6 +1,6 @@
 #include "historicalEvent.hpp"
 
-HistoricalEvent::HistoricalEvent(float d, std::vector<std::string> msg)
+HistoricalEvent::HistoricalEvent(long d, std::vector<std::string> msg)
 {
     std::vector<std::string>::iterator iter;
     for (iter = msg.begin(); iter != msg.end(); iter++)
@@ -17,7 +17,7 @@ std::vector<std::string> HistoricalEvent::getMessages()
     return messages;
 }
 
-float HistoricalEvent::getDate()
+long HistoricalEvent::getDate()
 {
     return date;
 }
